Ensure JSON schemas are in schema submodules
