+++ betty/model/__init__.py
@@ -12,7 +12,7 @@ from betty.json.linked_data import (
     JsonLdObject,
     LinkedDataDumpableWithSchemaJsonLdObject,
 )
-from betty.json.schema import Array, JsonSchemaReference, Null, OneOf, String
+from betty.json.schema import JsonSchemaReference, String
 from betty.locale.localizable import Localizable, _
 from betty.locale.localizer import DEFAULT_LOCALIZER
 from betty.mutability import Mutable
@@ -184,48 +184,3 @@ def persistent_id(entity_or_id: Entity | str, /) -> bool:
 
 
 _EntityT = TypeVar("_EntityT", bound=Entity)
-
-
-class ToZeroOrOneSchema(OneOf):
-    """
-    A schema for a to-zero-or-one entity association.
-    """
-
-    def __init__(self, *, title: str | None = None, description: str | None = None):
-        super().__init__(
-            String(
-                title=title or "Optional associate entity",
-                description=description
-                or "An optional reference to an associate entity's JSON resource",
-                format=String.Format.URI,
-            ),
-            Null(),
-        )
-
-
-class ToOneSchema(String):
-    """
-    A schema for a to-one entity association.
-    """
-
-    def __init__(self, *, title: str | None = None, description: str | None = None):
-        super().__init__(
-            title=title or "Associate entity",
-            description=description
-            or "A reference to an associate entity's JSON resource",
-            format=String.Format.URI,
-        )
-
-
-class ToManySchema(Array):
-    """
-    A schema for a to-many entity association.
-    """
-
-    def __init__(self, *, title: str | None = None, description: str | None = None):
-        super().__init__(
-            ToOneSchema(),
-            title=title or "Associate entities",
-            description=description
-            or "References to associate entities' JSON resources",
-        )
