+++ betty/tests/locale/localizable/test___init__.py
@@ -1,6 +1,5 @@
-from collections.abc import Callable, Iterable, Sequence
+from collections.abc import Callable, Sequence
 from gettext import NullTranslations
-from typing import cast
 
 import pytest
 from typing_extensions import override
@@ -21,7 +20,6 @@ from betty.locale.localizable import (
     ShorthandStaticTranslations,
     StaticTranslations,
     StaticTranslationsMapping,
-    StaticTranslationsSchema,
     UnorderedList,
     do_you_mean,
     ensure_localizable,
@@ -30,7 +28,6 @@ from betty.locale.localizable import (
 from betty.locale.localizer import DEFAULT_LOCALIZER, Localizer
 from betty.serde.dump import Dump, DumpMapping
 from betty.test_utils.json.linked_data import assert_dumps_linked_data
-from betty.test_utils.json.schema import SchemaTestBase, SchemaTestBaseSut
 
 
 class TestStaticTranslations:
@@ -207,40 +204,6 @@ class TestStaticTranslations:
         sut = StaticTranslations(translations, required=False)
         actual = await assert_dumps_linked_data(sut)
         assert actual == expected
-
-
-class TestStaticTranslationsSchema(SchemaTestBase):
-    @staticmethod
-    def _sut_params() -> Iterable[SchemaTestBaseSut]:
-        valid_datas: Sequence[Dump] = [
-            {DEFAULT_LOCALE: "Hello, world!"},
-            {"nl": "Hallo, wereld!", "uk": "Привіт Світ!"},
-        ]
-        invalid_datas: Sequence[Dump] = [
-            True,
-            False,
-            None,
-            123,
-            [],
-            {DEFAULT_LOCALE: True},
-            {DEFAULT_LOCALE: False},
-            {DEFAULT_LOCALE: None},
-            {DEFAULT_LOCALE: 123},
-            {DEFAULT_LOCALE: []},
-            {DEFAULT_LOCALE: {}},
-        ]
-        return [
-            (
-                StaticTranslationsSchema(),
-                valid_datas,
-                invalid_datas,
-            ),
-        ]
-
-    @override
-    @pytest.fixture(params=_sut_params())
-    def sut(self, request: pytest.FixtureRequest) -> SchemaTestBaseSut:
-        return cast(SchemaTestBaseSut, request.param)
 
 
 class TestPlain:
