+++ betty/tests/model/test___init__.py
@@ -1,6 +1,4 @@
-from collections.abc import Iterable
 from pathlib import Path
-from typing import cast
 
 import pytest
 from typing_extensions import override
@@ -9,14 +7,10 @@ from betty.locale.localizable import CountablePlain, Plain
 from betty.model import (
     Entity,
     EntityDefinition,
-    ToManySchema,
-    ToOneSchema,
-    ToZeroOrOneSchema,
     persistent_id,
 )
 from betty.plugin import PluginDefinition
 from betty.test_utils.documentation import PluginDocumentationTestBase
-from betty.test_utils.json.schema import SchemaTestBase, SchemaTestBaseSut
 from betty.test_utils.plugin.classed import ClassedPluginDefinitionClassTestBase
 
 
@@ -51,62 +45,3 @@ class TestEntityDocumentation(PluginDocumentationTestBase[EntityDefinition]):
 )
 def test_persistent_id(expected: bool, entity: Entity) -> None:
     assert persistent_id(entity) == expected
-
-
-class TestToOneSchema(SchemaTestBase):
-    @staticmethod
-    def _sut_params() -> Iterable[SchemaTestBaseSut]:
-        return [
-            (
-                ToOneSchema(),
-                [
-                    "https://example.com",
-                ],
-                [True, False, None, 123, [], {}],
-            ),
-        ]
-
-    @override
-    @pytest.fixture(params=_sut_params())
-    def sut(self, request: pytest.FixtureRequest) -> SchemaTestBaseSut:
-        return cast(SchemaTestBaseSut, request.param)
-
-
-class TestToZeroOrOneSchema(SchemaTestBase):
-    @staticmethod
-    def _sut_params() -> Iterable[SchemaTestBaseSut]:
-        return [
-            (
-                ToZeroOrOneSchema(),
-                [
-                    "https://example.com",
-                    None,
-                ],
-                [True, False, 123, [], {}],
-            ),
-        ]
-
-    @override
-    @pytest.fixture(params=_sut_params())
-    def sut(self, request: pytest.FixtureRequest) -> SchemaTestBaseSut:
-        return cast(SchemaTestBaseSut, request.param)
-
-
-class TestToManySchema(SchemaTestBase):
-    @staticmethod
-    def _sut_params() -> Iterable[SchemaTestBaseSut]:
-        return [
-            (
-                ToManySchema(),
-                [
-                    [],
-                    ["https://example.com"],
-                ],
-                [True, False, None, "123", 123, {}],
-            ),
-        ]
-
-    @override
-    @pytest.fixture(params=_sut_params())
-    def sut(self, request: pytest.FixtureRequest) -> SchemaTestBaseSut:
-        return cast(SchemaTestBaseSut, request.param)
