+++ betty/tests/privacy/test___init__.py
@@ -1,14 +1,12 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, cast
+from typing import TYPE_CHECKING, Any
 
 import pytest
-from typing_extensions import override
 
 from betty.privacy import (
     HasPrivacy,
     Privacy,
-    PrivacySchema,
     is_private,
     is_public,
     merge_privacies,
@@ -16,12 +14,9 @@ from betty.privacy import (
     resolve_privacy,
 )
 from betty.test_utils.json.linked_data import assert_dumps_linked_data
-from betty.test_utils.json.schema import SchemaTestBase, SchemaTestBaseSut
 from betty.test_utils.privacy import DummyHasPrivacy
 
 if TYPE_CHECKING:
-    from collections.abc import Iterable
-
     from betty.serde.dump import Dump, DumpMapping
 
 
@@ -118,19 +113,6 @@ class TestHasPrivacy:
         self, expected: DumpMapping[Dump], sut: HasPrivacy
     ) -> None:
         assert await assert_dumps_linked_data(sut) == expected
-
-
-class TestPrivacySchema(SchemaTestBase):
-    @staticmethod
-    def _sut_params() -> Iterable[SchemaTestBaseSut]:
-        return [
-            (PrivacySchema(), [True, False], [None, 123, "abc", [], {}]),
-        ]
-
-    @override
-    @pytest.fixture(params=_sut_params())
-    def sut(self, request: pytest.FixtureRequest) -> SchemaTestBaseSut:
-        return cast(SchemaTestBaseSut, request.param)
 
 
 @pytest.mark.parametrize(
